requirements.txt
asgiref==3.7.2
certifi==2024.7.4
channels==4.1.0
charset-normalizer==3.3.2
distlib==0.3.9
Django==5.0.3
django-ckeditor
django-js-asset==2.2.0
django-storages==1.14.4
et_xmlfile==2.0.0
filelock==3.15.3
idna==3.7
markdown==3.6  # 추가
numpy==2.2.2
oauthlib==3.2.2
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.4.4
toml==0.10.2
tzdata==2024.1
urllib3==1.26.19
virtualenv==20.26.2
webencodings==0.5.1
websockets==9.1
whitenoise==6.7.0